requirements.txt
# openai
numpy==1.26.4
pydantic
python-dotenv
requests
transformers>=4.40.0
torch>=2.2.0
# accelerate>=0.29.0
# bitsandbytes
openrouter
pyngrok
pandas
llama-cpp-python>=0.2.23
streamlit>=1.28.0
sdgx
pyarrow>=10.0.0
sdv
typing